+++ Sources/Development/main.swift
@@ -11,71 +11,9 @@ var workDir: String {
 }
 #endif
 
-public final class S {
-}
-
 import Foundation
 
-public enum Result<T> {
-    case success(T)
-    case failure(ErrorProtocol)
-
-    public func extract() throws -> T {
-        switch self {
-        case .success(let val):
-            return val
-        case .failure(let e):
-            throw e
-        }
-    }
-}
-
-public enum AsyncPromiseError: ErrorProtocol {
-    case senderNotCalled
-    case unexpectedEmptyResult
-    case timedOut
-}
-
-public final class Sender<T> {
-    private var result: Result<T>? = .none
-    private let semaphore: DispatchSemaphore
-
-    private init(_ semaphore: DispatchSemaphore) {
-        self.semaphore = semaphore
-    }
-
-    public func send(_ value: T) {
-        // TODO: Fatal error or throw? It's REALLY convenient NOT to throw here. Should at least log warning
-        guard result == nil else { return }
-        result = .success(value)
-        semaphore.signal()
-    }
-
-    public func send(_ error: ErrorProtocol) {
-        guard result == nil else { return }
-        result = .failure(error)
-        semaphore.signal()
-    }
-}
-
-public func asyncPromise<T>(returning: T.Type = T.self, timeout: DispatchTime = .distantFuture, _ handler: (Sender<T>) throws -> Void) throws -> T {
-    let semaphore = DispatchSemaphore(value: 0)
-    let sender = Sender<T>(semaphore)
-    // Ok to call synchronously, since will still unblock semaphore
-    // TODO: Find a way to enforce sender is called, not calling will perpetually block w/ long timeout
-    try handler(sender)
-    // TODO: Expose timeout customization -- I think Foundation is missing initializer
-    let semaphoreResult = semaphore.wait(timeout: timeout)
-    switch semaphoreResult {
-    case .Success:
-        guard let result = sender.result else { throw AsyncPromiseError.senderNotCalled }
-        return try result.extract()
-    case .TimedOut:
-        throw AsyncPromiseError.timedOut
-    }
-}
-
-let int: Int = try asyncPromise { sender in
+let int: Int = try Promise.async { sender in
     let url = URL(string: "https://api.spotify.com/v1/search/?q=beyonce&type=artist")
     URLSession.shared().dataTask(with: url!) { data, response, error in
         print("Completed Task")
@@ -100,6 +38,13 @@ let 😀 = Response(status: .ok)
 
 app.get { request in
     return try app.view("welcome.html")
+}
+
+app.get("async") { request in
+    return try Promise.async { promise in
+        // do whatever backgrounding is necessary here.
+        promise.send(JSON.object([:]))
+    }
 }
 
 app.get("client-socket") { req in
